requirements.txt
backports-abc==0.5
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
enum34==1.1.9
funcsigs==1.0.2
future==0.18.2
futures==3.1.1
ipaddress==1.0.23
mock==3.0.5
psycopg2==2.7.5
py==1.8.1
pycparser==2.19
pytest==2.9.1
python-dotenv==0.11.0
python-telegram-bot==12.4.2
singledispatch==3.4.0.3
six==1.14.0
tornado==5.1.1
typing==3.7.4.1
